Move WildFly integration tests from ear module to integration-tests module and introduce Arquillian and JUnit 5 (resolves #470)
