+++ integration-tests/wildfly/jgroups.cli
