+++ integration-tests/wildfly/src/test/java/org/jboss/test/clusterbench/it/wildfly/NodeNameServletIT.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.jboss.test.clusterbench.ear;
+package org.jboss.test.clusterbench.it.wildfly;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -17,42 +17,46 @@ import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.util.EntityUtils;
+import org.jboss.arquillian.container.test.api.RunAsClient;
+import org.jboss.arquillian.junit5.ArquillianExtension;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 /**
- * Test for {@link org.jboss.test.clusterbench.web.debug.HttpResponseServlet}.
- *
  * @author Radoslav Husar
  */
-public class HttpResponseServletIT {
+@ExtendWith(ArquillianExtension.class)
+@RunAsClient
+public class NodeNameServletIT extends AbstractWildFlyIT {
 
     @Test
     public void test() throws Exception {
         try (CloseableHttpClient httpClient = HttpClients.createDefault()) {
-            HttpGet httpGet = new HttpGet("http://localhost:8080/clusterbench/http-response?code=200");
+            HttpGet httpGet = new HttpGet("http://localhost:8080/clusterbench/jboss-node-name");
 
             try (CloseableHttpResponse response = httpClient.execute(httpGet)) {
                 assertEquals(200, response.getStatusLine().getStatusCode());
 
                 String responseBody = EntityUtils.toString(response.getEntity());
-                assertTrue(responseBody.contains("HTTP Code was: 200"));
-                assertTrue(responseBody.contains("JVM route: " + DebugServletIT.JBOSS_NODE_NAME));
-                assertTrue(responseBody.contains("Session isNew: true"));
+                assertEquals(DebugServletIT.JBOSS_NODE_NAME, responseBody);
 
-                // Also ensure session is created
+                // Also ensure session is created with ?create=true
                 Optional<Header> header = Arrays.stream(response.getAllHeaders()).filter(h -> Arrays.stream(h.getElements()).anyMatch(e -> e.getName().equals("JSESSIONID"))).findAny();
-                assertTrue(header.isPresent());
+                Assertions.assertFalse(header.isPresent());
             }
 
-            httpGet = new HttpGet("http://localhost:8080/clusterbench/http-response?code=503");
+            httpGet = new HttpGet("http://localhost:8080/clusterbench/jboss-node-name?create=true");
 
             try (CloseableHttpResponse response = httpClient.execute(httpGet)) {
-                assertEquals(503, response.getStatusLine().getStatusCode());
+                assertEquals(200, response.getStatusLine().getStatusCode());
 
                 String responseBody = EntityUtils.toString(response.getEntity());
-                assertTrue(responseBody.contains("HTTP Code was: 503"));
-                assertTrue(responseBody.contains("JVM route: " + DebugServletIT.JBOSS_NODE_NAME));
-                assertTrue(responseBody.contains("Session isNew: false"));
+                assertEquals(DebugServletIT.JBOSS_NODE_NAME, responseBody);
+
+                // Also ensure session is created with ?create=true
+                Optional<Header> header = Arrays.stream(response.getAllHeaders()).filter(h -> Arrays.stream(h.getElements()).anyMatch(e -> e.getName().equals("JSESSIONID"))).findAny();
+                assertTrue(header.isPresent());
             }
         }
     }
